move ptr package to top level pkg
ptr is a generic package not specific to the server.
