+++ mcp-server/pkg/tools/events/events.go
@@ -4,6 +4,7 @@ package events
 import (
 	"fmt"
 
+	"github.com/chronosphereio/mcp-server/pkg/ptr"
 	"github.com/go-openapi/strfmt"
 	"github.com/mark3labs/mcp-go/mcp"
 	"go.uber.org/zap"
@@ -11,7 +12,6 @@ import (
 	"github.com/chronosphereio/mcp-server/generated/dataunstable/dataunstable/data_unstable"
 	"github.com/chronosphereio/mcp-server/generated/dataunstable/models"
 	"github.com/chronosphereio/mcp-server/mcp-server/pkg/client"
-	"github.com/chronosphereio/mcp-server/mcp-server/pkg/ptr"
 	"github.com/chronosphereio/mcp-server/mcp-server/pkg/tools"
 	"github.com/chronosphereio/mcp-server/mcp-server/pkg/tools/pkg/params"
 )
