+++ tools/cmd/mcpgen/tmpl/entity_tool.go.tmpl
@@ -8,9 +8,9 @@ import (
 
 	"github.com/chronosphereio/mcp-server/generated/configv1/configv1/{{ .Entity.PkgName }}"
 	"github.com/chronosphereio/mcp-server/mcp-server/pkg/client"
-	"github.com/chronosphereio/mcp-server/mcp-server/pkg/ptr"
 	"github.com/chronosphereio/mcp-server/mcp-server/pkg/tools"
 	"github.com/chronosphereio/mcp-server/mcp-server/pkg/tools/pkg/params"
+	"github.com/chronosphereio/mcp-server/pkg/ptr"
 )
 
 {{ $entity := .Entity }}
